AudioDxe: Fix DevicePath not to contain padding

+++ Staging/AudioDxe/HdaController/HdaController.c
@@ -488,7 +488,7 @@ HdaControllerScanCodecs (
       HdaIoDevicePathNode.Header.Length[0]  = (UINT8)(sizeof (EFI_HDA_IO_DEVICE_PATH));
       HdaIoDevicePathNode.Header.Length[1]  = (UINT8)((sizeof (EFI_HDA_IO_DEVICE_PATH)) >> 8);
       HdaIoDevicePathNode.Guid              = gEfiHdaIoDevicePathGuid;
-      HdaIoDevicePathNode.Address           = (UINT8) Index;
+      HdaIoDevicePathNode.Address           = Index;
       HdaControllerDev->HdaIoChildren[Index].DevicePath = AppendDevicePathNode (HdaControllerDev->DevicePath, (EFI_DEVICE_PATH_PROTOCOL*)&HdaIoDevicePathNode);
       if (HdaControllerDev->HdaIoChildren[Index].DevicePath == NULL) {
         Status = EFI_INVALID_PARAMETER;
